Move countParticles to update methods
imgParticles should generally be kept current, and must be current for updateForward since it checks for particle overflow.  It no longer needs to be called by particleMerge (if it ever did), since at the moment particleMerge does not use it - it was only there so imgParticles would be updated once per update step. Although the dissipative update method does not refer to imgParticles it was added for consistency. In conclusion, imgParticles is only needed by update methods which utilize it. Soon particleMerge will be rewritten to use it.