Rearrange Pixeldust constructor, add transparentToWhite method
Remove initialization of particles[] and imgParticles[] from constructor.
Now they must be called separately, but this allows passing in particles.

Add transparentToWhite() method to deal with transparent png being
treated as black and spawning particles under image based init,
and in generation of imgParticles.

This commit pairs with the next one.