Require a trigger (not just a click) to start a person
Move most of draw() into run(). Abandon noLoop as control mechanism.
Were using noLoop() but we need to test some condition in draw() and
even with noLoop() draw is run once but we're not ready for it yet.
Changed sketch size since display window is >100 square even if sketch
size is 1, but need it to be bigger to click.