Add displayPixelsMasked and rename display to displayPixels
Mask out regions correponding to lighter pixels in original image. Works by using the inverse of the original image's pixel brightness as alpha for this frame's image (higher brightness in the reference image  results in more transparency for the final image) . A parameter allows transparency to be boosted linearly, to maintain visibility in fully white areas.