Improve maxAcceleration lower limit
By setting a smaller lower limit on maxAcceleration the net effect is to lower the maxVelocity at start of next phase. This causes a noticeably slower dissipation. Note with this way of setting maxVelocity, the lower limit of 0 causes a pause, and 1 seems to blow apart too fast, though likely frameRate dependent.