Bug fix: simulation begins in fall state, by changing isComplete logic
The previous commit changed the else if condition in draw() from
else if (isReady == 0) to
else if (sim != null)
which led to a bug resulting from a race condition.

Fix involved initializing isComplete of -1 and testing isComplete in the
else if clause in run().

In certain non-deterministic cases there would be a moment in begin()
when sim was created but before isComplete was first set to 0.
Then if(isComplete == 0) would fail and the else if(sim != null) would
pass and falling was observed when the simulation started.

With the same conditions simply moving the assignments of isComplete and
isReady to the start of run() did not work since the if clause would pass
and attempt to run() but the first sim had not been initialized yet.
Worse than the initial falling, the else if clause would occasionally
throw a NullPointerException due to the delay in initializing sim.

Also note, moving isComplete to the top of begin() is probably a bad idea
generally due to this delay in initializing sim coupled with the
concurrency issue. This is all caused by the fact that the network trigger
directly calls begin() and thus it's run in a separate thread. A better
solution might involve triggering via a variable which then causes the
main loop to call begin(). Setting isComplete and isReady in clientEvent
might help also.

This bug was observed sometimes upon first network trigger. Also observed
in subsequent triggerings but oddly seemed less likely there..
Bug was only observed with clientEvent (not mousePressed) triggering,
since it can trigger rapidly and since network stuff involves separate
threads (Client implements Java Runnable interface) thus thus concurrency
issues.

Note: falling on begin was not previously observed because of
else if (isReady == 0) condition coupled with the fact that
isReady was initialized to 1.

Note: setting sim = null at the start of begin() can help with debugging
since it increases likelihood of a crash instead of weird concurrency bugs.