Bug fix: solve concurrency issues by calling begin() from draw()
Many possible inconsistencies arose from calling begin() from clientEvent().
This cleaner solution involves renaming isComplete to triggerState, moving
(re)start logic to draw(), and communicating through the triggerState
variable. Directly calling begin() from clientEvent leads to separate
threads of execution and all sorts of weirdness.

Now in the general case draw() is not executed while we are in an in
inconsistent state while in the process of changing over to the next sim.
Initializing isComplete to -1 (see previous commit #88b094b) handles the
start-up state before the first sim is initialized.

Note: this also fixes the bug where 2 audio files were playing at once.
Like the premature falling, that occurred when begin() was running in a
separate thread after being called by clientEvent. In certain cases when
a new sim had been created but before isComplete was reset to 0 the sim
would be run as if it was complete. This resulted in falling dynamics as
well as repeated testing of hasFallen which could lead to isReady being
set back to 1 immediately (if new particles had not been fully
initialized and passed the fall test). This allowed another sim to be
triggered.