Implement handoff of particles to next person
Overload the PixeldustSimulation constructor so that an existing
particles array may now be passed in.
Existing particles are passed along to new person/sim from the previous
one. Now an entirely new set of particles is created only for the first
person (no longer for the first image of all people).

If nextImage is called and there are no existing particles then
particleIndices will be empty for the duration of the function call.
Cleverly use this fact so that particles are only started on the bottom
edge if none have been passed in.
Now particles are spawned at existing positions even for the first image
when particles are passed in, as well as for subsequent images.