Draw to off-screen buffer and make fullScreen work
Previously particles were positioned relative to the display surface,
centered, and at the same scale as the originating image. This made
rescaling a hassle and so it was left undone. So the window did not resize
in windowed mode, and the image did not scale at all when running fullScreen.
We were dealing just in pixels. To overcome these issues we now draw
to an off-screen graphics context which can be easily scaled for display.

One major bummer is that FX2D does not work for off-screen rendering so
we must use another slower renderer!