Remove surplus particles immediately when starting a new person
Surplus particles are generally sent off into space, but this behavior
is undesirable when we are starting a new simulation. Now unassigned
particles carried over from a previous person are immediately removed.

For some strange reason trying to integrate in with kill loop causes
program to hang when starting new person:

    if (particleIndexes.size() > 0) {
      if (imgIndex == 0) {
        for (int i = particleIndexes.size()-1; i > -1; i--) {
          particles.remove(particleIndexes.get(i));
        }
      } else {
        for (int i = 0; i < particleIndexes.size(); i++) {
          particles.get(particleIndexes.get(i)).kill();
        }
      }
    }