Make fall() method work
This code was moved into sim to fix a but where falling was done in
reference to display window and not the sims internal coordinate system.
Moving it seemed the easiest solution, and making it a method is more
natural anyway.

Only the minimal number of changes were made to fix the bug but
maintain the same logic, but it could stand to be cleaned up sometime.
The main concern is that triggerState really does not belong in sim as
it's a broader property. It might be worth making haveFallen a sim
instance variable and simply returning that, so that main can handle the
trigger logic independently.